+++ python-scripts/total-code-192.168.1.206.py
@@ -7,22 +7,67 @@ gc.collect()
 import uasyncio as asyncio
 gc.collect()
 mqtt_client = None
-nodes_id = ["70e3b87f23e7e8","1d9c9acba02ea5","fb0ed2cf7f5388"]
-input_topics = ["topic4_node","topic8_node"]
-output_topics = ["topic2_node","topic3_node","topic5_node","topic6_node"]
+nodes_id = ["5c254a58f2cab4","fb0ed2cf7f5388","70e3b87f23e7e8","1d9c9acba02ea5","fb0ed2cf7f5388"]
+input_topics = ["topic8_node","topic4_node"]
+output_topics = ["topic4_node","topic2_node","topic3_node","topic5_node","topic6_node"]
+
+import dht
+gc.collect()
+import machine
+gc.collect()
+output_topics_5c254a58f2cab4 = ["topic4_node"]
+pin_5c254a58f2cab4 = 32
+interval_5c254a58f2cab4 = 5000
+repeat_5c254a58f2cab4 = True
+
+reference_timer_workaround = []
+
+def measure__5c254a58f2cab4(_):
+    d = dht.DHT22(machine.Pin(pin_5c254a58f2cab4))
+    d.measure()
+    temperature = d.temperature()
+    humidity = d.humidity()
+    results = dict(
+        payload=dict(
+            temperature=temperature,
+            humidity=humidity
+        ) 
+    )
+    loop = asyncio.get_event_loop()
+    loop.create_task(on_output(ujson.dumps(results), output_topics_5c254a58f2cab4))
+
+def stop_5c254a58f2cab4():
+    for timer in reference_timer_workaround:
+        timer.deinit()
+
+def exec_5c254a58f2cab4():
+    if repeat_5c254a58f2cab4:
+        timer = machine.Timer(-1)    
+        timer.init(period=interval_5c254a58f2cab4, mode=machine.Timer.PERIODIC, callback=measure__5c254a58f2cab4)
+        reference_timer_workaround.append(timer)
+    else: 
+        measure(None)
+    return
+
+input_topics_fb0ed2cf7f5388 = ["topic8_node"]
+output_topics_fb0ed2cf7f5388 = ["results"]
+
+def on_input_fb0ed2cf7f5388(topic, msg, retained):
+    loop = asyncio.get_event_loop()
+    loop.create_task(on_output(msg, output_topics_fb0ed2cf7f5388))
 
 import dht
 gc.collect()
 import machine
 gc.collect()
 output_topics_70e3b87f23e7e8 = ["topic2_node","topic3_node"]
-pin_70e3b87f23e7e8 = 14
+pin_70e3b87f23e7e8 = 32
 interval_70e3b87f23e7e8 = 5000
 repeat_70e3b87f23e7e8 = True
 
 reference_timer_workaround = []
 
-def measure(_):
+def measure__70e3b87f23e7e8(_):
     d = dht.DHT22(machine.Pin(pin_70e3b87f23e7e8))
     d.measure()
     temperature = d.temperature()
@@ -43,7 +88,7 @@ def stop_70e3b87f23e7e8():
 def exec_70e3b87f23e7e8():
     if repeat_70e3b87f23e7e8:
         timer = machine.Timer(-1)    
-        timer.init(period=interval_70e3b87f23e7e8, mode=machine.Timer.PERIODIC, callback=measure)
+        timer.init(period=interval_70e3b87f23e7e8, mode=machine.Timer.PERIODIC, callback=measure__70e3b87f23e7e8)
         reference_timer_workaround.append(timer)
     else: 
         measure(None)
@@ -93,7 +138,9 @@ def on_input_fb0ed2cf7f5388(topic, msg, retained):
 
 def on_input(topic, msg, retained):
     topic = topic.decode()
-    if topic in input_topics_1d9c9acba02ea5:
+    if topic in input_topics_fb0ed2cf7f5388:
+        on_input_fb0ed2cf7f5388(topic, msg, retained)
+    elif topic in input_topics_1d9c9acba02ea5:
         on_input_1d9c9acba02ea5(topic, msg, retained)
     elif topic in input_topics_fb0ed2cf7f5388:
         on_input_fb0ed2cf7f5388(topic, msg, retained)
